redis event store occ tests + make occ tests a generic spec to be shared between modules
